Moved MainActivity
