+++ app/src/main/java/org/radarcns/kafka/KafkaDataSubmitter.java
@@ -2,7 +2,7 @@ package org.radarcns.kafka;
 
 import org.radarcns.data.DataCache;
 import org.radarcns.data.DataHandler;
-import org.radarcns.empaticaE4.MainActivity;
+import org.radarcns.MainActivity;
 import org.radarcns.util.ListPool;
 import org.radarcns.data.Record;
 import org.radarcns.kafka.rest.ServerStatusListener;
