+++ app/src/main/res/layout/activity_overview.xml
@@ -5,7 +5,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context=".empaticaE4.MainActivity">
+    tools:context=".MainActivity">
     <HorizontalScrollView
         android:id="@+id/horizontalScroll"
         android:layout_width="fill_parent"
